requirements.txt
pyopencl
numpy
silx
opencv-python
pyyaml
matplotlib
statistics